make FakeSyncSessionStoreManager inherit from MemcachedSessionManager and purge duplicated code
