+++ core/src/test/java/com/predic8/membrane/core/interceptor/oauth2/client/SyncSMOAuth2R2Test.java
@@ -14,6 +14,7 @@
 package com.predic8.membrane.core.interceptor.oauth2.client;
 
 import com.predic8.membrane.core.interceptor.oauth2client.OAuth2Resource2Interceptor;
+import com.predic8.membrane.core.interceptor.session.FakeSyncSessionStoreManager;
 
 public class SyncSMOAuth2R2Test extends OAuth2ResourceTest {
     @Override
